+++ langchain/chains/conversation/base.py
@@ -3,10 +3,10 @@ from typing import Dict, List
 
 from pydantic import BaseModel, Extra, Field, root_validator
 
-from langchain.chains.conversation.memory import ConversationBufferMemory
 from langchain.chains.conversation.prompt import PROMPT
 from langchain.chains.llm import LLMChain
 from langchain.memory.base import Memory
+from langchain.memory.buffer import ConversationBufferMemory
 from langchain.prompts.base import BasePromptTemplate
