+++ langchain/memory/chat_memory.py
@@ -1,7 +1,9 @@
-from typing import List
+from typing import Any, Dict, List, Optional
 
 from pydantic import BaseModel, Field
 
+from langchain.memory.base import Memory
+from langchain.memory.utils import get_prompt_input_key
 from langchain.schema import AIMessage, BaseMessage, HumanMessage
 
 
@@ -16,3 +18,28 @@ class ChatMessageHistory(BaseModel):
 
     def clear(self) -> None:
         self.messages = []
+
+
+class ChatMemoryMixin(Memory):
+    chat_memory: ChatMessageHistory = Field(default_factory=ChatMessageHistory)
+    output_key: Optional[str] = None
+    input_key: Optional[str] = None
+
+    def save_context(self, inputs: Dict[str, Any], outputs: Dict[str, str]) -> None:
+        """Save context from this conversation to buffer."""
+        if self.input_key is None:
+            prompt_input_key = get_prompt_input_key(inputs, self.memory_variables)
+        else:
+            prompt_input_key = self.input_key
+        if self.output_key is None:
+            if len(outputs) != 1:
+                raise ValueError(f"One output key expected, got {outputs.keys()}")
+            output_key = list(outputs.keys())[0]
+        else:
+            output_key = self.output_key
+        self.chat_memory.add_user_message(inputs[prompt_input_key])
+        self.chat_memory.add_ai_message(outputs[output_key])
+
+    def clear(self) -> None:
+        """Clear memory contents."""
+        self.chat_memory.clear()
