+++ tests/unit_tests/chains/test_conversation.py
@@ -2,12 +2,10 @@
 import pytest
 
 from langchain.chains.conversation.base import ConversationChain
-from langchain.chains.conversation.memory import (
-    ConversationBufferMemory,
-    ConversationBufferWindowMemory,
-    ConversationSummaryMemory,
-)
 from langchain.memory.base import Memory
+from langchain.memory.buffer import ConversationBufferMemory
+from langchain.memory.buffer_window import ConversationBufferWindowMemory
+from langchain.memory.summary import ConversationSummaryMemory
 from langchain.prompts.prompt import PromptTemplate
 from tests.unit_tests.llms.fake_llm import FakeLLM
 
@@ -16,14 +14,14 @@ def test_memory_ai_prefix() -> None:
     """Test that ai_prefix in the memory component works."""
     memory = ConversationBufferMemory(memory_key="foo", ai_prefix="Assistant")
     memory.save_context({"input": "bar"}, {"output": "foo"})
-    assert memory.buffer == "\nHuman: bar\nAssistant: foo"
+    assert memory.buffer == "Human: bar\nAssistant: foo"
 
 
 def test_memory_human_prefix() -> None:
     """Test that human_prefix in the memory component works."""
     memory = ConversationBufferMemory(memory_key="foo", human_prefix="Friend")
     memory.save_context({"input": "bar"}, {"output": "foo"})
-    assert memory.buffer == "\nFriend: bar\nAI: foo"
+    assert memory.buffer == "Friend: bar\nAI: foo"
 
 
 def test_conversation_chain_works() -> None:
