+++ tests/unit_tests/chains/test_sequential.py
@@ -4,8 +4,9 @@ from typing import Dict, List
 import pytest
 from pydantic import BaseModel
 
-from langchain.chains.base import Chain, SimpleMemory
+from langchain.chains.base import Chain
 from langchain.chains.sequential import SequentialChain, SimpleSequentialChain
+from langchain.memory.simple import SimpleMemory
 
 
 class FakeChain(Chain, BaseModel):
